+++ Sources/Views/Shelf/ShelfMain/ShelfView.swift
@@ -940,3 +940,6 @@ struct ShelfView: View {
             }
         }
     }
+}
+
+